Fix "›" not dispaying correctly on some browsers
The "›" character was displaying as "â€º" on Firefox 49.0.1.

Console output follows: "The character encoding of the HTML document was
not declared. The document will render with garbled text in some browser
configurations if the document contains characters from outside the
US-ASCII range. The character encoding of the page must be declared in
the document or in the transfer protocol."